requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
redis==5.0.1
langchain==0.1.3
langchain-community==0.0.8
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
httpx==0.25.2
aioredis==2.0.1
ollama==0.1.25
pyyaml==6.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-multipart==0.0.6
websockets==12.0
